업로드 api 테스트 미성공

+++ Clone/src/main/java/com/example/clone/controller/UserController.java
@@ -20,6 +20,7 @@ public class UserController {
 
     private final UserService userService;
 
+
     @Autowired
     public UserController(UserService userService) {
         this.userService = userService;
@@ -57,8 +58,7 @@ public class UserController {
         String nickname = userDetails.getUser().getNickname();
         String profileImage = userDetails.getUser().getProfileImage();
         String address = userDetails.getUser().getAddress();
-        Long rate = userDetails.getUser().getRate();
-
+        float rate = userDetails.getUser().getRate();
 
 
         ResponseEntity.ok()
